backend/requirements.txt
fastapi
uvicorn
python-multipart
openai-whisper
opencv-python
numpy
google-generativeai
python-dotenv
groq
firebase-admin==6.4.0
pdfplumber==0.10.3
scikit-learn==1.3.2
pydantic==2.5.3
email-validator
requests
cachetools
urllib3
annotated-types
typing-extensions
